ast: Move some AST-specific code from main package into AST package
This is to prepare AST for tests.
